Add files via upload
I have declared a 3 string variables and 1 int to store user inputted names, city and age. I've declared 2 variables for name so entering 2 names will be possible, I have also checked if the input for age is of numeric value and if its not the program will ask for another input.