Add files via upload
I've made a simple quiz about countries and capitals, I used a string array for the countries and capitals and then a for loop to check if the user input matches the capital of the country in the same index in both arrays. The input is fully case sensitive so an input like "PaRiS" will be treated as an invalid input or incorrect answer.