Add files via upload
I've made a program that sets a default password and sets 5 maximum attempts, then asks the user to enter the password. If the user enters correct password he is welcomed to the secured area if he doesn't and exceeds the 5 maximum attempts he is informed that the authorities have been alerted.